Adds S3 upload utility function, cleans up users.js route file, updates upload route, adds dotenv update in app.js
